Update fullyUpgradableToComponent2 utility function deprecated lifecycle method list
