+++ test/component2_suggestors/component2_utilities_test.dart
@@ -298,7 +298,7 @@ void main() {
         });
 
         group(
-            'extends a base class and contains lifecycle methods that are all updated by codemods',
+            'extends a base class and contains deprecated lifecycle methods that are all updated by codemods',
             () {
           final input = '''
             @Component
@@ -311,6 +311,9 @@ void main() {
               
               @override
               componentDidUpdate(Map prevProps, Map prevState) {}
+              
+              @override
+              shouldComponentUpdate() {}
             }
           ''';
 
@@ -392,7 +395,7 @@ void main() {
               componentDidUpdate(Map prevProps, Map prevState) {}
               
               @override
-              componentWillUnmount() {}
+              componentWillUpdate() {}
             }
           ''';
