+++ test/component2_suggestors/componentdidupdate_migrator_test.dart
@@ -94,7 +94,7 @@ void componentDidUpdateTests({bool allowPartialUpgrades}) {
         );
       });
 
-      test('-- has lifecycle methods without codemods', () {
+      test('-- has deprecated lifecycle methods without codemods', () {
         testSuggestor(
           expectedPatchCount: allowPartialUpgrades ? 1 : 0,
           input: '''
@@ -105,7 +105,7 @@ void componentDidUpdateTests({bool allowPartialUpgrades}) {
               }
               
               @override
-              componentWillUnmount() {}
+              componentWillReceiveProps() {}
             }
           ''',
           expectedOutput: '''
@@ -116,7 +116,7 @@ void componentDidUpdateTests({bool allowPartialUpgrades}) {
               }
               
               @override
-              componentWillUnmount() {}
+              componentWillReceiveProps() {}
             }
           ''',
         );
